+++ src/draggable.js
@@ -14,13 +14,6 @@ const throttledDragOver = (callback, duration) => {
   }
 }
 
-const delayedNativeDragStart = (callback, delay) => {
-  return (event) => {
-    event.stopPropagation()
-    setTimeout(() => callback(event), delay)
-  }
-}
-
 const passiveFalse = { passive: false }
 
 const isTouch = 'ontouchstart' in window
@@ -114,7 +107,7 @@ export default class Draggable extends EventEmitter {
     this._dragStart = (event) => this.dragStart(event)
     this._dragMove = (event) => this.dragMove(event)
     this._dragEnd = (event) => this.dragEnd(event)
-    this._nativeDragStart = delayedNativeDragStart((event) => this.nativeDragStart(event), this.nativeDragStartDelay)
+    this._nativeDragStart = (event) => this.nativeDragStart(event)
     this._nativeDragOver = throttledDragOver((event) => this.nativeDragOver(event), this.dragOverThrottleDuration)
     this._nativeDragEnd = (event) => this.nativeDragEnd(event)
     this._nativeDrop = (event) => this.nativeDrop(event)
@@ -561,10 +554,6 @@ export default class Draggable extends EventEmitter {
 
   get shouldRemoveZeroTranslate() {
     return this.options.shouldRemoveZeroTranslate || false
-  }
-
-  get nativeDragStartDelay() {
-    return this.options.nativeDragStartDelay || 0
   }
 
   get touchDraggingThreshold() {